[FIX] Field type

+++ software_knowledge_base_website_polling/models/swkb_installation_poll.py
@@ -46,7 +46,7 @@ class SWKBInstallationPoll(models.Model):
         group_operator='avg',
     )
 
-    timeout = fields.Integer(
+    timeout = fields.Float(
         string='Timeout',
     )
